[breaking] Huge overhaul of configuration types
Our configuration was very much written in the mindset of working
with text-only data such as that received by a modem's stdout output.
The problem with this is that not all data is actually textual or
parsable. We might want to trigger callbacks on a directory's contents
changing or on some other event taking place. Maybe the user wants to
trigger callbacks given some form of telemetry we haven't even thought
of.

This commit moves telemetry-specific callbacks to be function
parameters to the initialization of that telemetry handler. In other
words, instead of MapviewConfig taking, as a parameter, a list of
"ParsedPacketCallback"s, we pass those to whatever telemetry handler
produces something that needs to be parsed (such as modemStdout).

Then our MapviewConfig simplifes to dealing with:
- How do we get data into the system and (at the same time) what do we
  do with that data?
- What do we do immediately after that data comes in, before its
  callbacks get triggered?

Those two questions are then answered by the two fields of
MapviewConfig.

Signed-off-by: Ricky Elrod <ricky@elrod.me>